Revert "rsvalidate: Add compare-last-frame action for visual regression testing"
This reverts commit f57293989c4fc0bd3e30cdd4dc90adaffacff7b7.
