+++ utils/validate/Cargo.toml
@@ -14,8 +14,7 @@ gst-base.workspace = true
 gst-video.workspace = true
 gst-validate.workspace = true
 rqrr = { version = "0.7", default-features = false, features = ["img"] }
-image = { version = "0.25", default-features = false, features = ["png"] }
-image-compare = "0.4"
+image = { version = "0.25", default-features = false }
 
 [lib]
 name = "gstrsvalidate"
