+++ utils/validate/src/lib.rs
@@ -2,22 +2,10 @@
 
 use gst::glib;
 
-mod utils;
 mod check_last_frame_qrcode;
-mod compare_last_frame;
 
 fn plugin_init(plugin: &gst::Plugin) -> Result<(), glib::BoolError> {
     if let Err(err) = check_last_frame_qrcode::register_validate_actions(plugin) {
-        gst::warning!(
-            gst::CAT_RUST,
-            "Failed to register validate actions: {}",
-            err
-        );
-
-        return Err(err);
-    }
-
-    if let Err(err) = compare_last_frame::register_validate_actions(plugin) {
         gst::warning!(
             gst::CAT_RUST,
             "Failed to register validate actions: {}",
